added dockerfile
